:sparkles: add InjectionManager
For method overriding
